requirements.txt
mysql-connector-python
tabulate
colorama